refactor(LottoUseCase) : Controller 의존성을 LottoUseCase로 변경
