+++ src/main/java/lotto/Application.java
@@ -4,6 +4,7 @@ import lotto.adapter.inbound.InputView;
 import lotto.adapter.inbound.LottoController;
 import lotto.adapter.inbound.OutputView;
 import lotto.adapter.outbound.RandomNumberGenerator;
+import lotto.application.port.input.LottoUseCase;
 import lotto.application.service.LottoService;
 import lotto.application.port.output.NumberGenerator;
 
@@ -11,12 +12,12 @@ public class Application {
     public static void main(String[] args) {
         NumberGenerator numberGenerator = new RandomNumberGenerator();
 
-        LottoService lottoService = new LottoService(numberGenerator);
+        LottoUseCase lottoUseCase = new LottoService(numberGenerator);
 
         InputView inputView = new InputView();
         OutputView outputView = new OutputView();
 
-        LottoController lottoController = new LottoController(inputView, outputView, lottoService);
+        LottoController lottoController = new LottoController(inputView, outputView, lottoUseCase);
 
         lottoController.run();
     }
